Fix review comments and move secret command to root level
