+++ import-export-cli/cmd/mi/mi.go
@@ -25,7 +25,6 @@ import (
 	miDeactivateCmd "github.com/wso2/product-apim-tooling/import-export-cli/cmd/mi/deactivate"
 	miDeleteCmd "github.com/wso2/product-apim-tooling/import-export-cli/cmd/mi/delete"
 	miGetCmd "github.com/wso2/product-apim-tooling/import-export-cli/cmd/mi/get"
-	miSecretCmd "github.com/wso2/product-apim-tooling/import-export-cli/cmd/mi/secret"
 	miUpdateCmd "github.com/wso2/product-apim-tooling/import-export-cli/cmd/mi/update"
 	"github.com/wso2/product-apim-tooling/import-export-cli/utils"
 )
@@ -53,5 +52,4 @@ func init() {
 	MICmd.AddCommand(miUpdateCmd.UpdateCmd)
 	MICmd.AddCommand(miActivateCmd.ActivateCmd)
 	MICmd.AddCommand(miDeactivateCmd.DeactivateCmd)
-	MICmd.AddCommand(miSecretCmd.SecretCmd)
 }
